==> 02-processing.Rmd <==
```{r 01-load-data, echo=F}
load(here('output','01-proteomics.rdata'))
```

# Data processing

Before normalizing intensities, we first need to discard some hits because they
are not suitable for analysis:

 1. contaminants (*identifier starts with CON*)
 2. reversed sequences (*identifier starts with REV*)
 3. multiple hits (*peptides matching several proteins*)
 4. not enough unique peptides (*low confidence identification (< 2 peptides)*)

```{r filter-data}
# Filter hits
ms1=filter_hits(ms0)
```

## Processing quantified intensities

Following the aim of this experiment, we wish to compare the variation of protein 
expression between strains.

First, we compute the average intensities over the replicates.

Then, we discard proteins hits with missing values for XX strains or for YY replicates
types of media. 

Finally, we normalize the intensities by transforming to log2 and subtracting 
each sample's median. (cf formula below)


**Normalization method = Equalizing medians**

$$ norm.int_{sample} = log2(raw.int_{sample}) - log2(median_{sample}) $$
We filter out hits that contains missing values for either strains or replicates

```{r process-intensities}
# Process intensities
is.integer64 <- function(x) inherits(x, "integer64")

int_raw = ms1 %>% dplyr::select(uniprot=majority_protein_i_ds,starts_with("lfq")) %>% 
            #mutate(across(where(is.integer64), ~as.integer(.))) %>% # make sure to use numeric and not integer
            as.data.frame() %>% column_to_rownames('uniprot') %>%
            rename_with(everything(),.fn=gsub, pattern='lfq_intensity_', replacement='') %>%
            rename_with(.fn = str_to_upper)
            

int_md_norm = center_intensities(int_raw, center='median', tolog2=T) %>% 
          as.data.frame %>% rownames_to_column('uniprot') %>%
          as_tibble() %>%
          rename_with(everything(),.fn =gsub, pattern='lfq_intensity_', replacement='') %>% 
          rename_with(.fn = str_to_upper)
         
# TRYING DIFFERENT NORMALIZATIONS SCHEME (NOT READY)
INT_NORM = normalize_intensities(int_raw,df.group)
int_norm = int_md_norm 
#compare_exp(log2(int_raw),int_loess_norm, all=T)

long_int_norm = pivot_longer(int_norm  , cols=-UNIPROT, values_to='int2use',
                             names_to = c('strain','bioR','techR','day'),
                             names_pattern = "([^_]+)_([^_]+)_([^_]+)_([^_]+)") %>%
                group_by(strain,UNIPROT) %>% mutate(na_rep = sum.na(int2use))

# Intensities across strains  (default is average)
int_by_strain = pivot_wider(long_int_norm ,
              id_cols=UNIPROT,
              names_from = 'strain',
              names_glue = "{strain}",
              values_from = "int2use",
              values_fn=mean_
              )

na_by_strain = pivot_wider(long_int_norm,
              id_cols=UNIPROT,
              names_from = c('strain'),
              names_glue = "na_rep_{strain}",
              values_from = "int2use",
              values_fn=sum.na
              )

df_strains= left_join(int_by_strain,na_by_strain) %>% 
  rowwise %>%
  mutate( na_strains = sum.na(c_across(cols = starts_with('lfq_int'))) )

int_all = int_norm   %>% column_to_rownames('UNIPROT') %>% as.data.frame()

# Removing hits with missing values for more than one strain (using average intensities over replicates)
ms2= df_strains %>% filter(na_strains < 1)
int_filt_strains = ms2 %>% dplyr::select(-starts_with('na')) %>% column_to_rownames('UNIPROT') %>% as.data.frame()

# Save processed & normalized intensities data
write_rds(df_strains,here::here('output','02-normalized-intensities.rds'))
```


```{r 02-save-data, echo=F}
save.image(here('output','02-processed-data.rdata'))
```
