==> 04-QC.Rmd <==
```{r load-data-processed, echo=F}
chap_cur=04
#load(here('output',sprintf("%02d-processed-data.rdata",chap_cur-1)))
th_strip = theme(strip.background = element_blank(), strip.text = element_blank(),
                 panel.border = element_blank(),
                 panel.grid.major = element_line(size=0.5),
                 panel.grid.minor = element_line(size=0.25,linetype = '32'),
                 legend.position = 'none')

```

# Quality control {#qc}

First, we retrieve the LFQ-intensities of each protein hit (label-free quantitation).

```{r int-lfq}
# get lfq-peptide intensities (lfq=label-free quantitation) 
intensities = ms1 %>% dplyr::select(uniprot=majority_protein_i_ds,starts_with("lfq")) 
# Convert intensities to long format 
long_int_all = get_long_intensities(intensities) %>% mutate(int2use = log10_int)
```

## Total sample intensities

```{r tot_int, echo=F, fig.cap='Total intensities per sample',out.width='100%'}
# Sum of intensities per samples
tot_int_raw  = sum_intensities(int_raw)
tot_int_lfq  = sum_intensities(int_lfq,'lfq_intensity')
tot_int_norm = sum_intensities(int_norm)
tot_int_imp  = sum_intensities(int_bpca)

axis16 = theme(axis.title = element_text(size=24))
b1=draw_barplot_sumint(tot_int_raw,'Total intensity (raw)')+ scale_fill_metro() + axis16
b2=draw_barplot_sumint(tot_int_lfq,'Total intensity (lfq)')+ scale_fill_metro()+ axis16
b3=draw_barplot_sumint(tot_int_norm,'Total intensity (normalized)') + scale_fill_metro()+ axis16
b4=draw_barplot_sumint(tot_int_imp,'Total intensity (imputed)') + scale_fill_metro()+ axis16
QC1 = (b1| b2) / (b3 | b4) 
plot(QC1)

ggsave(QC1,filename =  here::here('plot','QC-total-intensities.png'), scale = 1.5)

```

## Boxplot of sample raw intensities

We want to inspect the distribution of peptide intensities between strains.
In addition, we will also observe in how many replicates each hit was quantified.

*The distribution of peptide intensities between all strains does not show strong differences of expression:*

```{r boxplot-int_ub, echo=F, fig.cap='Distribution of expression for ubiquitous hits (i.e. detected in all strains)'}

# Get intensities for ubiquitously detected hits between strains
long_intensities = long_int_all %>% 
  filter(ratio_na_rep<1) %>% # At least one peptide in one replicate from a each strain
  group_by(strain) %>% mutate(nprot = n_distinct(uniprot)) %>% 
  hablar::convert(n_na_rep=int(n_na_rep)) %>% 
  arrange(n_na_rep)

# Get median expression and # proteins per strain as a function of NAs in replicates
stats_per_strain=long_intensities %>% 
  group_by(strain=toupper(strain),n_na_rep=factor(n_na_rep)) %>% 
  summarize(nprot=n_distinct(uniprot),  MAX=max(int2use), MED=median_(int2use)) %>% 
  mutate( YMAX = MAX - 0.25*as.numeric(n_na_rep))

# Compare intensities of ubiquitous hits between strains
bp=ggplot(data=long_intensities,mapping=aes(y=int2use,x=toupper(strain))) +
  ggbeeswarm::geom_quasirandom(mapping = aes(group=n_na_rep,color=as.factor(nreplicates-n_na_rep)), na.rm=TRUE) +
  geom_violin(alpha=0.2, na.rm=TRUE,draw_quantiles = c(0.25,0.5,0.75)) + 
  geom_text(data=stats_per_strain,aes(label=nprot,color=nreplicates-n_na_rep,x=strain,y=YMAX),show.legend = F) +
  ylim(5,max_(stats_per_strain$MAX)) +
  ylab('Peptides intensities (log10)') + labs(x='Strains', colour='expressed in # replicates') +
  theme_pubclean() + grids(axis = 'y') + theme(axis.text = element_text(size=12)) +
  scale_color_canva(palette = "Pool party",limits=rev)
plot(bp)

ggsave(bp,filename =  here::here('plot','QC-strain-norm_log10_intensities.png'), scale = 0.9)

BY_STRAINS = group_by(long_intensities,strain=toupper(strain)) %>% 
             summarize(median_exp=median_(int2use)) %>%
             data.table::transpose(make.name = "strain", keep.names = "strain name") 

BY_REP = group_by(long_intensities, n_replicates=as.integer(4-n_na_rep)) %>% 
  summarize(median_exp=median_(int2use)) %>% 
  data.table::transpose(make.name = 'n_replicates',keep.names = '# replicates')

TAB_STRAINS_EXP = kbl( BY_STRAINS ,digits = 2) %>%
  kable_paper("striped", full_width = F) %>% kable_styling()

TAB_REP_EXP = kbl( BY_REP, digits=2 ) %>%
  kable_paper("striped", full_width = F) %>% kable_styling()  

```

The median peptide intensity are remarkably similar (~10^8^ or 7.9 in log10) between strains:
`r TAB_STRAINS_EXP`

*However, the peptide intensities can vary widely (>10-fold) when a hit is not detected across all replicates:*
`r TAB_REP_EXP`

**On average, peptide intensities are higher when a hit is found in more than one replicate**

## Count of missing hits per sample

The following barplot shows the range of missing hits per samples.

```{r missing-hit-samples,echo=F, fig.cap='Count of missing hits per samples',out.width='100%'}
count_NA_samples= pivot_longer(intensities,   -uniprot,
                           names_to = c('sample','strain','bio','tech','day'),
                           names_pattern = "(([^_]+)_([^_]+)_([^_]+)_([^_]+))",
                           names_prefix = 'lfq_intensity_',
                           values_to='int') %>% 
  group_by(sample) %>% 
  summarize(n_na=sum(is.na(int))) %>% 
  dplyr::left_join(df.group %>% mutate(sample=tolower(sample))) %>% 
  mutate(is_outlier = strain %in% c("CQC","CMP") & biorep=="R1")

mean_na_outliers = mean_(count_NA_samples %>% dplyr::filter(is_outlier) %>% pull(n_na)) %>% round()
mean_na = mean_(count_NA_samples %>% dplyr::filter(!is_outlier) %>% pull(n_na)) %>% round()

bp_na = ggplot(count_NA_samples,aes(x=sample,label=sample)) + 
  geom_col(aes(y=n_na,fill=strain,col=is_outlier)) + 
  geom_text(data=subset(count_NA_samples,!is_outlier),aes(y=0),angle=90,hjust=-0.1,size=2.5)+
  geom_text(data=subset(count_NA_samples,is_outlier),aes(y=0),angle=90,hjust=-0.1,fontface='bold',size=2.5)+
  ylab('# NAs') +
#  geom_hline(yintercept = mean_na_outliers,size=1) + 
  geom_hline(yintercept = mean_na,linetype=2,size=0.5) +
  scale_fill_metro() +
  scale_color_manual(values = c('TRUE'='black','FALSE'='white')) + 
  facet_wrap(~strain,scale='free_x',nrow = 2) +
  th_strip + 
  theme(axis.text.x = element_blank(),axis.ticks = element_blank())

plot(bp_na)
```

CQC and CMP are the two strains containing the most missing values. 

Particularly, the first biological replicate of those strains have twice more missing
protein intensities (`r mean_na_outliers` hits so 10-16% NAs) while the rest of the samples
have on average about only `r mean_na` hits missing (4% to 9% NAs).

We then look at hits quantified as a function of the number of strains in which 
at least one peptide was quantified. (non-ubiquitous hits)

```{r missing-hit, echo=F, fig.cap='Distribution of expression for hits not detected in all strains',out.width='100%'}
pep_missing = long_int_all %>%
  filter(ratio_na_rep==1) %>% 
  dplyr::select(-c(bio,tech)) %>% 
  distinct()

hit_missing = pep_missing$uniprot

bp_miss= long_int_all %>% arrange(n_na_strains) %>% 
         dplyr::filter(has_missing_strains) %>%
  ggplot(mapping=aes(y=int2use,x=8-n_na_strains)) +
  geom_violin(mapping=aes(fill=factor(n_na_strains)),
              alpha=0.2, na.rm=TRUE,draw_quantiles = c(0.25,0.5,0.75),color='black',show.legend = F) +
  geom_text(aes(label=nprot_miss),y=12,check_overlap = T,col='red') + ylim(5,12) +
  ylab('Peptides intensities (log10)') + labs(x='expressed across # strains') +
  scale_x_continuous(breaks=0:8) +
  theme_pubclean() + grids(axis = 'y') + theme(axis.text = element_text(size=12))
  
plot(bp_miss)

stat_per_miss = long_int_all %>%
  dplyr::select(n_na_strains,nprot,int2use) %>%
  mutate(n_strains= nstrains-n_na_strains) %>%
  group_by(n_strains) %>% mutate(md_exp = median_(int2use)) %>%
  ungroup() %>% dplyr::select(-int2use) %>%
  distinct() %>% 
  arrange(n_strains) %>% 
  mutate(
          f_exp = 100*nprot/sum(nprot),
          pc_exp = sprintf("%.1f%%",f_exp),
          pc_cum = sprintf("%.1f%%",cumsum(f_exp))
          ) %>%
  arrange(desc(n_strains)) %>%
  mutate(pc_cum_desc = sprintf("%.1f%%",cumsum(f_exp))) %>%
  dplyr::select(nprot,n_strains,md_exp,pc_exp,pc_cum,pc_cum_desc)


TAB_EXP_STRAIN = kbl(stat_per_miss,digits = 2, align = 'l',
                     col.names=c('# proteins','# expressing strains','median exp. (log10)','% quantified','cum. %','cum. % (decreasing)')) %>%
  kable_paper("striped", full_width = F) %>% kable_styling()

```

`r TAB_EXP_STRAIN`

**Proteins expressed across less strains have lower median peptide intensity.**
Nevertheless, almost two thirds (*66.5%*) of the protein hits were detected in all strains.
About 86% protein hits were expressed in the majority of the strains (at least 5 out of 8 strains).

## Missing hits across samples

Furthermore, we can check whether the proportion of missing values is equally 
distributed among all samples.

```{r na-samples, echo=F, fig.cap='Number of quantified intensities across samples', out.width='100%'}
# Count hits with NA in samples/groups
na_int= intensities %>% 
        group_by(uniprot) %>%
        summarize(across(all_of(paste0("lfq_intensity_",samples)), is.na ) )
colnames(na_int) = c('uniprot',samples)

p_upset = ComplexUpset::upset(
  data=as.data.frame(1-na_int[,-1]),
  intersect=samples,
  name='intensity quantified', 
  min_degree=5, min_size=10,
  sort_sets=F, keep_empty_groups=F, wrap=TRUE,
  width_ratio=0.1, 
  base_annotations=list('Number of hits'=ComplexUpset::intersection_size(
                              text_colors=c(on_background='red', on_bar='yellow'),
                              text=list(angle=0,hjust=0.45,size=5)) +
    annotate(geom='text', x=Inf, y=Inf, label=paste('Total hits:', nrow(na_int)), vjust=1,hjust=1 )
  ), themes = ComplexUpset::upset_default_themes(text=element_text(size=26))
)

plot(p_upset)
# w1200px =  unit(1200,'px')
ggsave(p_upset, path=here('plot'),
        filename = 'upset-samples-combination-nacount.pdf', 
        device = 'pdf',scale = 2, dpi=300,  width = 1200, units='px')

```


## Coefficient of variations

Coefficient of variations correspond to the percent of variance relative to the mean.
The following boxplots show how variable protein expression is, across strains
or across biological replicates, before and after normalization.

```{r boxplot-cv, echo=F, fig.cap='Coefficient of variations (%) across samples for each normalization method'}

CV_raw = calculate_cv(long_int_all,by_bio = F) 
box_cv_raw = draw_boxcv(CV_raw,by_bio = F)
plot(box_cv_raw)
#CV_norm = calculate_cv(long_int_norm,by_bio = F)
# box_cv_norm = draw_boxcv(CV_norm,by_bio = F)
# plot(box_cv_raw / box_cv_norm)

CV_norm = INT_NORM@ner@avgcvmem %>% as_tibble() %>%
          add_column(strain=unique(df.group$strain) ) %>%
          group_by(strain) %>%
          pivot_longer(-strain, names_to = 'normalization', values_to='CV' ) %>% 
          group_by(normalization) %>% 
          mutate(outlier = ifelse(is.outlier(CV,thr = 0.25,coef = 1.5), strain, NA))


box_cv_norm = ggplot(CV_norm, aes(x=fct_relevel(normalization,'log2'),y=CV)) + 
  geom_boxplot(fill='gray',outlier.shape=21, outlier.size=0.5, outlier.fill = NA) + xlab('Normalization method') + ylab('%CV samples')  +
  geom_text(aes(label = outlier), na.rm = TRUE, vjust = -0.5, size=2.5)
plot(box_cv_norm)

ggsave(box_cv_norm,
       filename = here::here('plot','QC-boxplot-normalization.png'),scale=1.1)
#show_table_cv(CV_raw,by_bio = F)
#show_table_cv(CV_raw,by_bio = T)
#show_table_cv(CV_norm,by_bio = F)
#show_table_cv(CV_norm,by_bio = T)
```

## Expression distributions

To highlight the strength of normalization, we also show the density distribution of 
expression before and after normalization using each of the following normalization methods:

`r TAB_NORM`

```{r exp_density,fig.height=8, fig.width=5}
dd=draw_normalization_density(int_raw,int_lfq,df.group) 
plot(dd)
ggsave(dd, filename=here::here('plot','QC-normalization-density.png'), scale = 1.5)
```


### Distribution of imputed missing values

Missing values imputed with Bayesian PCA estimator are not modifying the distribution
of normalized protein expression across strains. 

The difference of average protein expression of imputed versus normalized intensities is well centered on 0 for most strains.

The outliers strains in overall protein expression (CMP and CQC) that contained
more missing values are slighlty shifted to negative values (i.e. lower expression).


```{r imp_vs_norm, fig.height=6, fig.width=12,out.width='100%'}
long_bpca = int_bpca %>% 
          rownames_to_column('uniprot') %>% 
          pivot_longer(-uniprot, 
                       names_to = c('strain','bio','tech','day'),
                       names_pattern = "([^_]+)_([^_]+)_([^_]+)_([^_]+)",
                       values_to = 'int_imp') %>% 
          group_by(uniprot,strain) %>% mutate(avg_imp = mean_(int_imp)) 


long_int = pivot_longer(df_int_norm , cols=-uniprot, values_to='int2use',
                             names_to = c('strain','bio','tech','day'),
                             names_pattern = "([^_]+)_([^_]+)_([^_]+)_([^_]+)") %>%
                group_by(strain,uniprot) %>%
                mutate(na_rep = sum.na(int2use),  avg_norm = mean_(int2use)) %>%
                dplyr::rename(int_norm = int2use)


df_exp = left_join(long_bpca,long_int) %>% 
         mutate(is_imputed = na_rep > 0) %>% 
         filter(is_imputed) %>% 
         mutate(delta_imp = avg_imp - avg_norm) 
n_exp = left_join(long_bpca,long_int) %>% 
        mutate(is_imputed = na_rep > 0) %>%
        group_by(strain) %>%  
        summarize(
           n_uni = n_distinct(uniprot[!is_imputed]),
           n_imp = n_distinct(uniprot[is_imputed]),
           n_toshow = sprintf("N=%s\nNA=%s",n_uni,n_imp)
        )


p_imp = ggplot(df_exp) + 
  geom_density(aes(x=int_imp,color=strain,fill=strain),bw=0.2) +
  geom_density(aes(x=int_norm),alpha=0.6,col=NA,fill='white',bw=0.2) +
  geom_rug(aes(x=int_imp,color=strain),alpha=.1) +
  geom_text(data=n_exp,aes(label=n_toshow,color=strain),x=Inf,y=Inf,size=3.5,
            hjust=1.1,vjust=1.2) +
  geom_text(data=n_exp,aes(label=strain,color=strain),x=-Inf,y=Inf,size=4,
            fontface='bold', hjust=-0.1,vjust=1.5) +
  scale_color_metro() +  scale_fill_metro() + 
  xlab("Distribution of average intensities") +
  facet_wrap(~strain,nrow = 2) + 
  th_strip

ggsave(p_imp, filename = here::here('plot','density-imputed-vs-norm.png'),scale=1.2)
ggsave(p_imp, filename = here::here('plot','density-imputed-vs-norm.pdf'),scale=1.2)
plot(p_imp)
```

### Difference of imputed and normalized intensities

```{r delta_imp_norm, out.width='100%'}
d_exp = df_exp %>%
        group_by(strain) %>%  
        summarize( mu = mean_(delta_imp), sigma = sd_(delta_imp))

p_delta = ggplot(df_exp) + 
  geom_density(aes(x=delta_imp,color=strain,fill=strain),bw=0.2,col=NA,alpha=0.6) +
  geom_rug(aes(x=delta_imp,color=strain),alpha=.1) +
  geom_text(data=n_exp,aes(label=n_toshow,color=strain),x=Inf,y=Inf,size=3.5,
            hjust=1.1,vjust=1.2) +
  geom_text(data=n_exp,aes(label=strain,color=strain),x=-Inf,y=Inf,size=4,
            fontface='bold', hjust=-0.1,vjust=1.5) +
  geom_linerange(d_exp, mapping=aes(y=0.7,xmin = mu-sigma, xmax = mu+sigma),color='black') + 
  geom_segment(d_exp, mapping=aes(x = mu, xend = mu, y=0,yend=Inf),color='black') + 
  scale_color_metro() +  scale_fill_metro() + 
  xlab("Distribution of (imputed-normalized) intensities") +
  facet_wrap(~strain,nrow = 4) + 
  th_strip


ggsave(p_delta, filename = here::here('plot','density-delta-imputed-vs-norm.png'),scale=1.2)
ggsave(p_delta, filename = here::here('plot','density-delta-imputed-vs-norm.pdf'),scale=1.2)

plot(p_delta)
```

## Compare all-vs-all expression

### Scatterplots

```{r scatterplot-all, fig.path='plot/', fig.show='hold', warning=F, fig.height=15, fig.width=15, out.width='100%'}
# scatterplots between all samples
scmat_all=draw_scatterplots(datain=int_all)
#print(scmat_all)

ggsave(scmat_all,filename=here::here('plot','scatterplot-samples-intensities.pdf'),width = 1200, units='px')

knitr::include_graphics(here('plot','scatterplot-samples-intensities.pdf'),auto_pdf = T)
```

### Heatmap correlation 

```{r heatmap-all, fig.path='plot/', fig.show='hold', warning=F, fig.height=15, fig.width=15, out.width='100%' }
# heatmap correlation
cs_all=compute_samples_correlation(int_all)
COR_RANGE = range( cs_all[ upper.tri(cs_all) ] )

by_sample = df.group %>% column_to_rownames("sample")
hm_all=draw_heatmap_samples(mcor = cs_all,df.group=by_sample,col.group = col.group)
```


The heatmap correlations of all samples show the high correlation of expression 
between replicates and between most strains:
`r sprintf("[%.3f - %.3f]",COR_RANGE[1],COR_RANGE[2])`

The expression from the 1st biological replicates of strains CQC and CMP seem
slighlty less correlated to the other samples.

### Principle Component Analysis

Finally, the PCA  reveals the distance between each sample.

```{r pca-all, fig.path='plot/', fig.show='hold', warning=F, out.width='100%'}
int_all_scaled = scale(int_all,center=T, scale = T)
make_pca(na.omit(int_all_scaled), with_labels=T,col_by_group=1:4)
```


```{r save-data-qc, echo=F}
save.image(here('output',sprintf("%02d-quality-control.rdata",chap_cur)))
```
