==> 05-strain-comparison.Rmd <==
```{r load-data-qc, echo=F}
chap_cur=05
#load(here('output',sprintf("%02d-quality-control.rdata",chap_cur-1)))
```

# Strains expression comparison

## Scatterplots 

Initially, we can look at the scatterplots of intensities all-versus-all samples.

```{r scatterplot-strains, fig.path='plot/', fig.show='hold', out.width='100%',warning=F}
# scatterplots between strains
scmat_strains=draw_scatterplots(datain=int_filt_strains)
print(scmat_strains)
```

## Heatmap Correlation

We then compute spearman rank correlations of intensities between all samples.

```{r heatmap-strains}
# heatmap correlation
cs_strains=compute_samples_correlation(int_filt_strains) %>% as.matrix
COR_RANGE = range( cs_strains[ upper.tri(cs_strains) ] )
by_strains= df.group[,c('strain'),drop=F] %>% distinct() %>% as.data.frame()
rownames(by_strains) = by_strains$strain
hm_strains=draw_heatmap_samples(mcor = cs_strains, df.group=by_strains, col.group = col.group,k=3)
```

The heatmap correlations show the high correlation of expression between strains:
`r sprintf("[%.3f - %.3f]",COR_RANGE[1],COR_RANGE[2])`


## Principle Component Analysis

Finally, the PCA  reveals the distance between each sample.

```{r pca-strains}
int_scaled_strains = scale(int_filt_strains,center=T, scale = T)
make_pca(na.omit(int_scaled_strains), with_labels=F,col_by_group=1:2)
```


```{r save-data-comparison, echo=F}
save.image(here('output',sprintf("%02d-strain-comparison.rdata",chap_cur)))
```
